+++ bid-svc/src/main/java/com/ebidding/bid/Repository/BidRankRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface BidRankRepository extends JpaRepository<BidRank, String> {
-    Optional<BidRank> findBybidId(int inputId);
+    Optional<BidRank> findBybidId(Long inputId);
 }
