+++ bid-svc/src/main/java/com/ebidding/bid/Repository/BidRepository.java
@@ -3,7 +3,7 @@ package com.ebidding.bid.Repository;
 import com.ebidding.bid.domain.Bid;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BidRepository extends JpaRepository<Bid, Integer> {
+public interface BidRepository extends JpaRepository<Bid, Long> {
     // Optional<Bid> findById(int inputId);
 
 }
